+++ examples/example3/main.go
@@ -1,4 +1,4 @@
-// Copyright 2012-2014 The GoSNMP Authors. All rights reserved.  Use of this
+// Copyright 2012 The GoSNMP Authors. All rights reserved.  Use of this
 // source code is governed by a BSD-style license that can be found in the
 // LICENSE file.
 
@@ -7,8 +7,9 @@ package main
 import (
 	"fmt"
 	"log"
+	"time"
 
-	g "github.com/soniah/gosnmp"
+	g "github.com/gosnmp/gosnmp"
 )
 
 func main() {
@@ -20,6 +21,7 @@ func main() {
 		Version:       g.Version3,
 		SecurityModel: g.UserSecurityModel,
 		MsgFlags:      g.AuthPriv,
+		Timeout:       time.Duration(30) * time.Second,
 		SecurityParameters: &g.UsmSecurityParameters{UserName: "user",
 			AuthenticationProtocol:   g.SHA,
 			AuthenticationPassphrase: "password",