+++ v3_usm.go
@@ -1,4 +1,4 @@
-// Copyright 2012-2020 The GoSNMP Authors. All rights reserved.  Use of this
+// Copyright 2012 The GoSNMP Authors. All rights reserved.  Use of this
 // source code is governed by a BSD-style license that can be found in the
 // LICENSE file.
 
@@ -15,11 +15,17 @@ import (
 	"crypto/cipher"
 	"crypto/des" //nolint:gosec
 	"crypto/hmac"
+	"crypto/md5" //nolint:gosec
 	crand "crypto/rand"
+	"crypto/sha1"     //nolint:gosec
+	_ "crypto/sha256" // Register hash function #4 (SHA224), #5 (SHA256)
+	_ "crypto/sha512" // Register hash function #6 (SHA384), #7 (SHA512)
 	"encoding/binary"
+	"encoding/hex"
 	"errors"
 	"fmt"
 	"hash"
+	"strings"
 	"sync"
 	"sync/atomic"
 )
@@ -40,6 +46,7 @@ const (
 
 //go:generate stringer -type=SnmpV3AuthProtocol
 
+// HashType maps the AuthProtocol's hash type to an actual crypto.Hash object.
 func (authProtocol SnmpV3AuthProtocol) HashType() crypto.Hash {
 	switch authProtocol {
 	default:
@@ -123,6 +130,7 @@ const (
 
 // UsmSecurityParameters is an implementation of SnmpV3SecurityParameters for the UserSecurityModel
 type UsmSecurityParameters struct {
+	mu sync.Mutex
 	// localAESSalt must be 64bit aligned to use with atomic operations.
 	localAESSalt uint64
 	localDESSalt uint32
@@ -146,13 +154,69 @@ type UsmSecurityParameters struct {
 	Logger Logger
 }
 
+// Description logs authentication paramater information to the provided GoSNMP Logger
+func (sp *UsmSecurityParameters) Description() string {
+	var sb strings.Builder
+	sb.WriteString("user=")
+	sb.WriteString(sp.UserName)
+
+	sb.WriteString(",engine=(")
+	sb.WriteString(hex.EncodeToString([]byte(sp.AuthoritativeEngineID)))
+	// sb.WriteString(sp.AuthoritativeEngineID)
+	sb.WriteString(")")
+
+	switch sp.AuthenticationProtocol {
+	case NoAuth:
+		sb.WriteString(",auth=noauth")
+	case MD5:
+		sb.WriteString(",auth=md5")
+	case SHA:
+		sb.WriteString(",auth=sha")
+	case SHA224:
+		sb.WriteString(",auth=sha224")
+	case SHA256:
+		sb.WriteString(",auth=sha256")
+	case SHA384:
+		sb.WriteString(",auth=sha384")
+	case SHA512:
+		sb.WriteString(",auth=sha512")
+	}
+	sb.WriteString(",authPass=")
+	sb.WriteString(sp.AuthenticationPassphrase)
+
+	switch sp.PrivacyProtocol {
+	case NoPriv:
+		sb.WriteString(",priv=NoPriv")
+	case DES:
+		sb.WriteString(",priv=DES")
+	case AES:
+		sb.WriteString(",priv=AES")
+	case AES192:
+		sb.WriteString(",priv=AES192")
+	case AES256:
+		sb.WriteString(",priv=AES256")
+	case AES192C:
+		sb.WriteString(",priv=AES192C")
+	case AES256C:
+		sb.WriteString(",priv=AES256C")
+	}
+	sb.WriteString(",privPass=")
+	sb.WriteString(sp.PrivacyPassphrase)
+
+	return sb.String()
+}
+
 // Log logs security paramater information to the provided GoSNMP Logger
 func (sp *UsmSecurityParameters) Log() {
+	sp.mu.Lock()
+	defer sp.mu.Unlock()
 	sp.Logger.Printf("SECURITY PARAMETERS:%+v", sp)
 }
 
 // Copy method for UsmSecurityParameters used to copy a SnmpV3SecurityParameters without knowing it's implementation
 func (sp *UsmSecurityParameters) Copy() SnmpV3SecurityParameters {
+	sp.mu.Lock()
+	defer sp.mu.Unlock()
 	return &UsmSecurityParameters{AuthoritativeEngineID: sp.AuthoritativeEngineID,
 		AuthoritativeEngineBoots: sp.AuthoritativeEngineBoots,
 		AuthoritativeEngineTime:  sp.AuthoritativeEngineTime,
@@ -175,6 +239,13 @@ func (sp *UsmSecurityParameters) getDefaultContextEngineID() string {
 	return sp.AuthoritativeEngineID
 }
 func (sp *UsmSecurityParameters) initSecurityKeys() error {
+	sp.mu.Lock()
+	defer sp.mu.Unlock()
+
+	return sp.initSecurityKeysNoLock()
+}
+
+func (sp *UsmSecurityParameters) initSecurityKeysNoLock() error {
 	var err error
 
 	if sp.AuthenticationProtocol > NoAuth && len(sp.SecretKey) == 0 {
@@ -189,7 +260,7 @@ func (sp *UsmSecurityParameters) initSecurityKeys() error {
 		switch sp.PrivacyProtocol {
 		// Changed: The Output of SHA1 is a 20 octets array, therefore for AES128 (16 octets) either key extension algorithm can be used.
 		case AES, AES192, AES256, AES192C, AES256C:
-			//Use abstract AES key localization algorithms
+			// Use abstract AES key localization algorithms.
 			sp.PrivacyKey, err = genlocalPrivKey(sp.PrivacyProtocol, sp.AuthenticationProtocol,
 				sp.PrivacyPassphrase,
 				sp.AuthoritativeEngineID)
@@ -212,40 +283,21 @@ func (sp *UsmSecurityParameters) setSecurityParameters(in SnmpV3SecurityParamete
 	var insp *UsmSecurityParameters
 	var err error
 
+	sp.mu.Lock()
+	defer sp.mu.Unlock()
+
 	if insp, err = castUsmSecParams(in); err != nil {
 		return err
 	}
 
 	if sp.AuthoritativeEngineID != insp.AuthoritativeEngineID {
 		sp.AuthoritativeEngineID = insp.AuthoritativeEngineID
+		sp.SecretKey = nil
+		sp.PrivacyKey = nil
 
-		if sp.AuthenticationProtocol > NoAuth && len(sp.SecretKey) == 0 {
-			sp.SecretKey, err = genlocalkey(sp.AuthenticationProtocol,
-				sp.AuthenticationPassphrase,
-				sp.AuthoritativeEngineID)
-			if err != nil {
-				return err
-			}
-		}
-		if sp.PrivacyProtocol > NoPriv && len(sp.PrivacyKey) == 0 {
-			switch sp.PrivacyProtocol {
-			// Changed: The Output of SHA1 is a 20 octets array, therefore for AES128 (16 octets) either key extension algorithm can be used.
-			case AES, AES192, AES256, AES192C, AES256C:
-				//Use abstract AES key localization algorithms
-				sp.PrivacyKey, err = genlocalPrivKey(sp.PrivacyProtocol, sp.AuthenticationProtocol,
-					sp.PrivacyPassphrase,
-					sp.AuthoritativeEngineID)
-				if err != nil {
-					return err
-				}
-			default:
-				sp.PrivacyKey, err = genlocalkey(sp.AuthenticationProtocol,
-					sp.PrivacyPassphrase,
-					sp.AuthoritativeEngineID)
-				if err != nil {
-					return err
-				}
-			}
+		err = sp.initSecurityKeysNoLock()
+		if err != nil {
+			return err
 		}
 	}
 	sp.AuthoritativeEngineBoots = insp.AuthoritativeEngineBoots
@@ -301,14 +353,14 @@ func (sp *UsmSecurityParameters) init(log Logger) error {
 		salt := make([]byte, 8)
 		_, err = crand.Read(salt)
 		if err != nil {
-			return fmt.Errorf("error creating a cryptographically secure salt: %s", err.Error())
+			return fmt.Errorf("error creating a cryptographically secure salt: %w", err)
 		}
 		sp.localAESSalt = binary.BigEndian.Uint64(salt)
 	case DES:
 		salt := make([]byte, 4)
 		_, err = crand.Read(salt)
 		if err != nil {
-			return fmt.Errorf("error creating a cryptographically secure salt: %s", err.Error())
+			return fmt.Errorf("error creating a cryptographically secure salt: %w", err)
 		}
 		sp.localDESSalt = binary.BigEndian.Uint32(salt)
 	}
@@ -497,6 +549,8 @@ func genlocalkey(authProtocol SnmpV3AuthProtocol, passphrase string, engineID st
 // http://tools.ietf.org/html/rfc2574#section-8.1.1.1
 // localDESSalt needs to be incremented on every packet.
 func (sp *UsmSecurityParameters) usmAllocateNewSalt() interface{} {
+	sp.mu.Lock()
+	defer sp.mu.Unlock()
 	var newSalt interface{}
 
 	switch sp.PrivacyProtocol {
@@ -509,6 +563,8 @@ func (sp *UsmSecurityParameters) usmAllocateNewSalt() interface{} {
 }
 
 func (sp *UsmSecurityParameters) usmSetSalt(newSalt interface{}) error {
+	sp.mu.Lock()
+	defer sp.mu.Unlock()
 	switch sp.PrivacyProtocol {
 	case AES, AES192, AES256, AES192C, AES256C:
 		aesSalt, ok := newSalt.(uint64)
@@ -565,31 +621,102 @@ func (sp *UsmSecurityParameters) discoveryRequired() *SnmpPacket {
 	return nil
 }
 
-func (sp *UsmSecurityParameters) calcPacketDigest(packet []byte) []byte {
-	var mac hash.Hash
+func (sp *UsmSecurityParameters) calcPacketDigest(packet []byte) ([]byte, error) {
+	return calcPacketDigest(packet, sp)
+}
 
-	switch sp.AuthenticationProtocol {
-	default:
-		mac = hmac.New(crypto.MD5.New, sp.SecretKey)
-	case SHA:
-		mac = hmac.New(crypto.SHA1.New, sp.SecretKey)
-	case SHA224:
-		mac = hmac.New(crypto.SHA224.New, sp.SecretKey)
-	case SHA256:
-		mac = hmac.New(crypto.SHA256.New, sp.SecretKey)
-	case SHA384:
-		mac = hmac.New(crypto.SHA384.New, sp.SecretKey)
-	case SHA512:
-		mac = hmac.New(crypto.SHA512.New, sp.SecretKey)
+// calcPacketDigest calculate authenticate digest for incoming messages (TRAP or
+// INFORM).
+// Support MD5, SHA1, SHA224, SHA256, SHA384, SHA512 protocols
+func calcPacketDigest(packetBytes []byte, secParams *UsmSecurityParameters) ([]byte, error) {
+	var digest []byte
+	var err error
+
+	switch secParams.AuthenticationProtocol {
+	case MD5, SHA:
+		digest, err = digestRFC3414(
+			secParams.AuthenticationProtocol,
+			packetBytes,
+			secParams.SecretKey)
+	case SHA224, SHA256, SHA384, SHA512:
+		digest, err = digestRFC7860(
+			secParams.AuthenticationProtocol,
+			packetBytes,
+			secParams.SecretKey)
 	}
 
-	_, _ = mac.Write(packet)
+	return digest, err
+}
+
+// digestRFC7860 calculate digest for incoming messages using HMAC-SHA2 protcols
+// according to RFC7860 4.2.2
+func digestRFC7860(h SnmpV3AuthProtocol, packet []byte, authKey []byte) ([]byte, error) {
+	mac := hmac.New(h.HashType().New, authKey)
+	_, err := mac.Write(packet)
+	if err != nil {
+		return []byte{}, err
+	}
 	msgDigest := mac.Sum(nil)
-	return msgDigest
+	return msgDigest, nil
+}
+
+// digestRFC3414 calculate digest for incoming messages using MD5 or SHA1
+// according to RFC3414 6.3.2 and 7.3.2
+func digestRFC3414(h SnmpV3AuthProtocol, packet []byte, authKey []byte) ([]byte, error) {
+	var extkey [64]byte
+	var err error
+	var k1, k2 [64]byte
+	var h1, h2 hash.Hash
+
+	copy(extkey[:], authKey)
+
+	switch h {
+	case MD5:
+		h1 = md5.New() //nolint:gosec
+		h2 = md5.New() //nolint:gosec
+	case SHA:
+		h1 = sha1.New() //nolint:gosec
+		h2 = sha1.New() //nolint:gosec
+	}
+
+	for i := 0; i < 64; i++ {
+		k1[i] = extkey[i] ^ 0x36
+		k2[i] = extkey[i] ^ 0x5c
+	}
+
+	_, err = h1.Write(k1[:])
+	if err != nil {
+		return []byte{}, err
+	}
+
+	_, err = h1.Write(packet)
+	if err != nil {
+		return []byte{}, err
+	}
+
+	d1 := h1.Sum(nil)
+
+	_, err = h2.Write(k2[:])
+	if err != nil {
+		return []byte{}, err
+	}
+
+	_, err = h2.Write(d1)
+	if err != nil {
+		return []byte{}, err
+	}
+
+	return h2.Sum(nil)[:12], nil
 }
 
 func (sp *UsmSecurityParameters) authenticate(packet []byte) error {
-	msgDigest := sp.calcPacketDigest(packet)
+	var msgDigest []byte
+	var err error
+
+	if msgDigest, err = sp.calcPacketDigest(packet); err != nil {
+		return err
+	}
+
 	idx := bytes.Index(packet, macVarbinds[sp.AuthenticationProtocol])
 
 	if idx < 0 {
@@ -602,6 +729,7 @@ func (sp *UsmSecurityParameters) authenticate(packet []byte) error {
 
 // determine whether a message is authentic
 func (sp *UsmSecurityParameters) isAuthentic(packetBytes []byte, packet *SnmpPacket) (bool, error) {
+	var msgDigest []byte
 	var packetSecParams *UsmSecurityParameters
 	var err error
 
@@ -609,8 +737,9 @@ func (sp *UsmSecurityParameters) isAuthentic(packetBytes []byte, packet *SnmpPac
 		return false, err
 	}
 	// TODO: investigate call chain to determine if this is really the best spot for this
-
-	msgDigest := sp.calcPacketDigest(packetBytes)
+	if msgDigest, err = calcPacketDigest(packetBytes, packetSecParams); err != nil {
+		return false, err
+	}
 
 	for k, v := range []byte(packetSecParams.AuthenticationParameters) {
 		if msgDigest[k] != v {
@@ -672,10 +801,13 @@ func (sp *UsmSecurityParameters) encryptPacket(scopedPdu []byte) ([]byte, error)
 }
 
 func (sp *UsmSecurityParameters) decryptPacket(packet []byte, cursor int) ([]byte, error) {
-	_, cursorTmp := parseLength(packet[cursor:])
+	_, cursorTmp, err := parseLength(packet[cursor:])
+	if err != nil {
+		return nil, err
+	}
 	cursorTmp += cursor
 	if cursorTmp > len(packet) {
-		return nil, fmt.Errorf("error decrypting ScopedPDU: truncated packet")
+		return nil, errors.New("error decrypting ScopedPDU: truncated packet")
 	}
 
 	switch sp.PrivacyProtocol {
@@ -696,7 +828,7 @@ func (sp *UsmSecurityParameters) decryptPacket(packet []byte, cursor int) ([]byt
 		packet = packet[:cursor+len(plaintext)]
 	default:
 		if len(packet[cursorTmp:])%des.BlockSize != 0 {
-			return nil, fmt.Errorf("error decrypting ScopedPDU: not multiple of des block size")
+			return nil, errors.New("error decrypting ScopedPDU: not multiple of des block size")
 		}
 		preiv := sp.PrivacyKey[8:]
 		var iv [8]byte
@@ -729,12 +861,18 @@ func (sp *UsmSecurityParameters) marshal(flags SnmpV3MsgFlags) ([]byte, error) {
 	buf.WriteString(sp.AuthoritativeEngineID)
 
 	// msgAuthoritativeEngineBoots
-	msgAuthoritativeEngineBoots := marshalUvarInt(sp.AuthoritativeEngineBoots)
+	msgAuthoritativeEngineBoots, err := marshalUint32(sp.AuthoritativeEngineBoots)
+	if err != nil {
+		return nil, err
+	}
 	buf.Write([]byte{byte(Integer), byte(len(msgAuthoritativeEngineBoots))})
 	buf.Write(msgAuthoritativeEngineBoots)
 
 	// msgAuthoritativeEngineTime
-	msgAuthoritativeEngineTime := marshalUvarInt(sp.AuthoritativeEngineTime)
+	msgAuthoritativeEngineTime, err := marshalUint32(sp.AuthoritativeEngineTime)
+	if err != nil {
+		return nil, err
+	}
 	buf.Write([]byte{byte(Integer), byte(len(msgAuthoritativeEngineTime))})
 	buf.Write(msgAuthoritativeEngineTime)
 
@@ -776,55 +914,39 @@ func (sp *UsmSecurityParameters) unmarshal(flags SnmpV3MsgFlags, packet []byte, 
 	var err error
 
 	if PDUType(packet[cursor]) != Sequence {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model parameters")
+		return 0, errors.New("error parsing SNMPV3 User Security Model parameters")
 	}
-	_, cursorTmp := parseLength(packet[cursor:])
+	_, cursorTmp, err := parseLength(packet[cursor:])
+	if err != nil {
+		return 0, err
+	}
 	cursor += cursorTmp
 	if cursorTmp > len(packet) {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model parameters: truncated packet")
+		return 0, errors.New("error parsing SNMPV3 User Security Model parameters: truncated packet")
 	}
 
 	rawMsgAuthoritativeEngineID, count, err := parseRawField(sp.Logger, packet[cursor:], "msgAuthoritativeEngineID")
 	if err != nil {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthoritativeEngineID: %s", err.Error())
+		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthoritativeEngineID: %w", err)
 	}
 	cursor += count
 	if AuthoritativeEngineID, ok := rawMsgAuthoritativeEngineID.(string); ok {
 		if sp.AuthoritativeEngineID != AuthoritativeEngineID {
 			sp.AuthoritativeEngineID = AuthoritativeEngineID
-			sp.Logger.Printf("Parsed authoritativeEngineID %s", AuthoritativeEngineID)
-			if sp.AuthenticationProtocol > NoAuth && len(sp.SecretKey) == 0 {
-				sp.SecretKey, err = genlocalkey(sp.AuthenticationProtocol,
-					sp.AuthenticationPassphrase,
-					sp.AuthoritativeEngineID)
-				if err != nil {
-					return 0, err
-				}
-			}
-			if sp.PrivacyProtocol > NoPriv && len(sp.PrivacyKey) == 0 {
-				switch sp.PrivacyProtocol {
-				case AES, AES192, AES256, AES192C, AES256C:
-					sp.PrivacyKey, err = genlocalPrivKey(sp.PrivacyProtocol, sp.AuthenticationProtocol,
-						sp.PrivacyPassphrase,
-						sp.AuthoritativeEngineID)
-					if err != nil {
-						return 0, err
-					}
-				default:
-					sp.PrivacyKey, err = genlocalkey(sp.AuthenticationProtocol,
-						sp.PrivacyPassphrase,
-						sp.AuthoritativeEngineID)
-					if err != nil {
-						return 0, err
-					}
-				}
+			sp.SecretKey = nil
+			sp.PrivacyKey = nil
+
+			sp.Logger.Printf("Parsed authoritativeEngineID %0x", []byte(AuthoritativeEngineID))
+			err = sp.initSecurityKeysNoLock()
+			if err != nil {
+				return 0, err
 			}
 		}
 	}
 
 	rawMsgAuthoritativeEngineBoots, count, err := parseRawField(sp.Logger, packet[cursor:], "msgAuthoritativeEngineBoots")
 	if err != nil {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthoritativeEngineBoots: %s", err.Error())
+		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthoritativeEngineBoots: %w", err)
 	}
 	cursor += count
 	if AuthoritativeEngineBoots, ok := rawMsgAuthoritativeEngineBoots.(int); ok {
@@ -834,7 +956,7 @@ func (sp *UsmSecurityParameters) unmarshal(flags SnmpV3MsgFlags, packet []byte, 
 
 	rawMsgAuthoritativeEngineTime, count, err := parseRawField(sp.Logger, packet[cursor:], "msgAuthoritativeEngineTime")
 	if err != nil {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthoritativeEngineTime: %s", err.Error())
+		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthoritativeEngineTime: %w", err)
 	}
 	cursor += count
 	if AuthoritativeEngineTime, ok := rawMsgAuthoritativeEngineTime.(int); ok {
@@ -844,7 +966,7 @@ func (sp *UsmSecurityParameters) unmarshal(flags SnmpV3MsgFlags, packet []byte, 
 
 	rawMsgUserName, count, err := parseRawField(sp.Logger, packet[cursor:], "msgUserName")
 	if err != nil {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgUserName: %s", err.Error())
+		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgUserName: %w", err)
 	}
 	cursor += count
 	if msgUserName, ok := rawMsgUserName.(string); ok {
@@ -854,7 +976,7 @@ func (sp *UsmSecurityParameters) unmarshal(flags SnmpV3MsgFlags, packet []byte, 
 
 	rawMsgAuthParameters, count, err := parseRawField(sp.Logger, packet[cursor:], "msgAuthenticationParameters")
 	if err != nil {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthenticationParameters: %s", err.Error())
+		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgAuthenticationParameters: %w", err)
 	}
 	if msgAuthenticationParameters, ok := rawMsgAuthParameters.(string); ok {
 		sp.AuthenticationParameters = msgAuthenticationParameters
@@ -862,18 +984,28 @@ func (sp *UsmSecurityParameters) unmarshal(flags SnmpV3MsgFlags, packet []byte, 
 	}
 	// blank msgAuthenticationParameters to prepare for authentication check later
 	if flags&AuthNoPriv > 0 {
+		// In case if the authentication protocol is not configured or set to NoAuth, then the packet cannot
+		// be processed further
+		if sp.AuthenticationProtocol <= NoAuth {
+			return 0, errors.New("error parsing SNMPv3 User Security Model: authentication parameters are not configured to parse incoming authenticated message")
+		}
 		copy(packet[cursor+2:cursor+len(macVarbinds[sp.AuthenticationProtocol])], macVarbinds[sp.AuthenticationProtocol][2:])
 	}
 	cursor += count
 
 	rawMsgPrivacyParameters, count, err := parseRawField(sp.Logger, packet[cursor:], "msgPrivacyParameters")
 	if err != nil {
-		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgPrivacyParameters: %s", err.Error())
+		return 0, fmt.Errorf("error parsing SNMPV3 User Security Model msgPrivacyParameters: %w", err)
 	}
 	cursor += count
 	if msgPrivacyParameters, ok := rawMsgPrivacyParameters.(string); ok {
 		sp.PrivacyParameters = []byte(msgPrivacyParameters)
 		sp.Logger.Printf("Parsed privacyParameters %s", msgPrivacyParameters)
+		if flags&AuthPriv >= AuthPriv {
+			if sp.PrivacyProtocol <= NoPriv {
+				return 0, errors.New("error parsing SNMPv3 User Security Model: privacy parameters are not configured to parse incoming encrypted message")
+			}
+		}
 	}
 
 	return cursor, nil